pyproject.toml
[tool.poetry]
name = "repl-nix-discoversync"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
requests = "^2.32.3"
jinja2 = "^3.1.4"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
pyjwt = "^2.9.0"
pydantic = "^2.9.0"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
